Feature: Create a start weight that can be use against a current weight
Goal is to compare the start weight based on team name of each player with the current weight for each player. 

The idea is that the player wouldn't have the same start weight for each team. Also when joining a team, the current weight is used as the start weight.